% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{dqRNGkind}
\alias{dqRNGkind}
\title{Choose RNGkind for DQRNG}
\usage{
dqRNGkind(kind, normal_kind = "ignored")
}
\arguments{
\item{kind}{string specifying the RNG (see details)}

\item{normal_kind}{ignored; included for compatibility with \code{\link{RNGkind}}}
}
\description{
Original in \url{https://github.com/daqana/dqrng/blob/master/src/dqrng.cpp}
}
\seealso{
\code{\link[dqrng]{dqRNGkind}}
}
