% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{dqset_seed}
\alias{dqset_seed}
\title{Set the seed for the DQRNG}
\usage{
dqset_seed(seed, stream = NULL)
}
\arguments{
\item{seed}{integer scalar to seed the random number generator, or an integer vector of length 2 representing a 64-bit seed.}

\item{stream}{integer used for selecting the RNG stream; either a scalar or a vector of length 2}
}
\description{
Original in \url{https://github.com/daqana/dqrng/blob/master/src/dqrng.cpp}
}
\seealso{
\code{\link[dqrng]{dqset.seed}}
}
