% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rexp-alternative}
\alias{rexp-alternative}
\alias{Rcpp_rexp_RNGScope}
\alias{Rcpp_rexp_RCPPRNG}
\alias{Rcpp_rexp_DQRNG}
\title{Alternatives for \code{rexp}}
\usage{
Rcpp_rexp_RNGScope(n, rate = 1)

Rcpp_rexp_RCPPRNG(n, rate = 1)

Rcpp_rexp_DQRNG(n, rate = 1)
}
\arguments{
\item{n}{Number of samples}

\item{rate}{Rate parameter}
}
\description{
Alternatives for \code{rexp}
}
