% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rexp-Rcpp}
\alias{rexp-Rcpp}
\alias{Rcpp_rexp1}
\alias{Rcpp_rexp2}
\alias{Rcpp_rexp3}
\alias{Rcpp_rexp4}
\alias{Rcpp_rexp5}
\title{\code{rexp} implementations in Rcpp for benchmarking}
\usage{
Rcpp_rexp1(n, rate = 1)

Rcpp_rexp2(n, rate = 1)

Rcpp_rexp3(n, rate = 1)

Rcpp_rexp4(n, rate = 1)

Rcpp_rexp5(n, rate = 1)
}
\arguments{
\item{n}{number of observations. If \code{length(n) > 1}, the length
    is taken to be the number required.}

\item{rate}{vector of rates.}
}
\description{
These functions serve no real purpose except demonstrating
how to use the RNG Generators and providing some functions
for benchmarks.
}
